+++ packages/ast-tooling/tsconfig.json
@@ -4,5 +4,6 @@
 		"checkJs": false,
 		"isolatedDeclarations": true,
 		"declaration": true
-	}
+	},
+	"exclude": ["vitest.config.ts"]
 }
