+++ packages/ast-tooling/utils.ts
@@ -1,0 +1,34 @@
+// Sourced from `golden-fleece`
+// https://github.com/Rich-Harris/golden-fleece/blob/f2446f331640f325e13609ed99b74b6a45e755c2/src/patch.ts#L302
+export function guessIndentString(str: string): string {
+	const lines = str.split('\n');
+
+	let tabs = 0;
+	let spaces = 0;
+	let minSpaces = 8;
+
+	lines.forEach((line) => {
+		const match = /^(?: +|\t+)/.exec(line);
+		if (!match) return;
+
+		const whitespace = match[0];
+		if (whitespace.length === line.length) return;
+
+		if (whitespace[0] === '\t') {
+			tabs += 1;
+		} else {
+			spaces += 1;
+			if (whitespace.length > 1 && whitespace.length < minSpaces) {
+				minSpaces = whitespace.length;
+			}
+		}
+	});
+
+	if (spaces > tabs) {
+		let result = '';
+		while (minSpaces--) result += ' ';
+		return result;
+	} else {
+		return '\t';
+	}
+}
