+++ packages/ast-tooling/vitest.config.ts
@@ -1,0 +1,8 @@
+import { defineProject } from 'vitest/config';
+
+export default defineProject({
+	test: {
+		name: 'ast-tooling',
+		include: ['./tests/*.ts']
+	}
+});
